refactor: extract postgres backend accessors and repository wiring into mixins (#2050)
